LLVM.AST

Description

This module and descendants define AST data types to represent LLVM code. Note that these types are designed for fidelity rather than convenience - if the truth of what LLVM supports is less than pretty, so be it.

Documentation

data Module

http://llvm.org/docs/LangRef.html#module-structure

Constructors

Module

Fields

moduleName :: ShortByteString

moduleSourceFileName :: ShortByteString

moduleDataLayout :: Maybe DataLayout

a DataLayout, if specified, must match that of the eventual code generator

moduleTargetTriple :: Maybe ShortByteString

moduleDefinitions :: [Definition]

defaultModule :: Module

helper for making Modules

data Definition

Any thing which can be at the top level of a Module

Constructors

GlobalDefinition Global

TypeDefinition Name (Maybe Type)

MetadataNodeDefinition MetadataNodeID MDNode

NamedMetadataDefinition ShortByteString [MetadataNodeID]

ModuleInlineAssembly ByteString

FunctionAttributes GroupID [FunctionAttribute]

COMDAT ShortByteString SelectionKind

data Global

http://llvm.org/doxygen/classllvm_1_1GlobalValue.html

Constructors

GlobalVariable Name Linkage Visibility (Maybe StorageClass) (Maybe Model) (Maybe UnnamedAddr) Bool Type AddrSpace (Maybe Constant) (Maybe ShortByteString) (Maybe ShortByteString) Word32 [(ShortByteString, MDRef MDNode)]
http://llvm.org/docs/LangRef.html#global-variables

GlobalAlias Name Linkage Visibility (Maybe StorageClass) (Maybe Model) (Maybe UnnamedAddr) Type AddrSpace Constant
http://llvm.org/docs/LangRef.html#aliases

Function Linkage Visibility (Maybe StorageClass) CallingConvention [ParameterAttribute] Type Name ([Parameter], Bool) [Either GroupID FunctionAttribute] (Maybe ShortByteString) (Maybe ShortByteString) Word32 (Maybe ShortByteString) (Maybe Constant) [BasicBlock] (Maybe Constant) [(ShortByteString, MDRef MDNode)]
http://llvm.org/docs/LangRef.html#functions

globalVariableDefaults :: Global

helper for making GlobalVariables

globalAliasDefaults :: Global

helper for making GlobalAliass

functionDefaults :: Global

helper for making Functions

data UnnamedAddr

Constructors

LocalAddr

GlobalAddr

data Parameter

Parameters for Functions

Constructors

Parameter Type Name [ParameterAttribute]

data BasicBlock

http://llvm.org/doxygen/classllvm_1_1BasicBlock.html LLVM code in a function is a sequence of BasicBlocks each with a label, some instructions, and a terminator.

Constructors

BasicBlock Name [Named Instruction] (Named Terminator)

data Operand

An Operand is roughly that which is an argument to an Instruction

Constructors

LocalReference Type Name
%foo

ConstantOperand Constant
Constants include GlobalReference, for @foo

MetadataOperand Metadata

type CallableOperand = Either InlineAssembly Operand

The Call instruction is special: the callee can be inline assembly

data Metadata

http://llvm.org/docs/LangRef.html#metadata

Constructors

MDString ShortByteString
http://llvm.org/docs/doxygen/html/classllvm_1_1MDNode.html

MDNode (MDRef MDNode)
http://llvm.org/docs/doxygen/html/classllvm_1_1MDNode.html

MDValue Operand
http://llvm.org/docs/doxygen/html/classllvm_1_1ValueAsMetadata.html

newtype MetadataNodeID

A MetadataNodeID is a number for identifying a metadata node. Note this is different from "named metadata", which are represented with NamedMetadataDefinition.

Constructors

MetadataNodeID Word

data MDRef a

MDRef can either represent a reference to some piece of metadata or the metadata itself.

This is mainly useful for encoding cyclic metadata. Note that LLVM represents inline and non-inline nodes identically, so roundtripping the Haskell AST does not preserve whether a node was inline or not.

Constructors

MDRef MetadataNodeID

MDInline a

data MDNode

http://llvm.org/docs/LangRef.html#metadata

Constructors

MDTuple [Maybe Metadata]
Nothing represents null

DIExpression DIExpression

DIGlobalVariableExpression DIGlobalVariableExpression

DILocation DILocation

DIMacroNode DIMacroNode

DINode DINode

module LLVM.AST.Instruction

module LLVM.AST.Name

data Type

http://llvm.org/docs/LangRef.html#type-system

Constructors

VoidType
http://llvm.org/docs/LangRef.html#void-type

IntegerType
http://llvm.org/docs/LangRef.html#integer-type

Fields

typeBits :: Word32

PointerType
http://llvm.org/docs/LangRef.html#pointer-type

Fields

pointerReferent :: Type

pointerAddrSpace :: AddrSpace

FloatingPointType
http://llvm.org/docs/LangRef.html#floating-point-types

Fields

floatingPointType :: FloatingPointType

FunctionType
http://llvm.org/docs/LangRef.html#function-type

Fields

resultType :: Type

argumentTypes :: [Type]

isVarArg :: Bool

VectorType
http://llvm.org/docs/LangRef.html#vector-type

Fields

nVectorElements :: Word32

elementType :: Type

StructureType
http://llvm.org/docs/LangRef.html#structure-type

Fields

isPacked :: Bool

elementTypes :: [Type]

ArrayType
http://llvm.org/docs/LangRef.html#array-type

Fields

nArrayElements :: Word64

elementType :: Type

NamedTypeReference Name
http://llvm.org/docs/LangRef.html#opaque-structure-types

MetadataType
http://llvm.org/docs/LangRef.html#metadata-type

LabelType
http://llvm.org/docs/LangRef.html#label-type

TokenType
http://llvm.org/docs/LangRef.html#token-type

data FloatingPointType

LLVM supports some special formats floating point format. This type is to distinguish those format. Also see http://llvm.org/docs/LangRef.html#floating-point-types

Constructors

HalfFP
16-bit floating point value

FloatFP
32-bit floating point value

DoubleFP
64-bit floating point value

FP128FP
128-bit floating point value (112-bit mantissa)

X86_FP80FP
80-bit floating point value (X87)

PPC_FP128FP
128-bit floating point value (two 64-bits)

Overview

llvm-hs-pure defines the Haskell AST for representing an LLVM Module. For interacting with the LLVM C/C++ libraries and an overview of the various libraries in the llvm-hs ecosystem, take a look at the docs in the LLVM module in llvm-hs.

In addition to constructing the LLVM AST manually, there is also a monadic IRBuilder interface in IRBuilder. The IRBuilder will take care of generating fresh names automatically and generally reduces the verbosity of using the AST directly. Using RecursiveDo/mdo, it is also capable of handling forward references automatically.

Constructing the AST for an LLVM module