LLVM.AST.Attribute

Description

Module to allow importing Attribute distinctly qualified. Before LLVM 3.5, the attributes which could be used on functions and those which could be used on parameters were disjoint. In LLVM 3.5, two attributes (readonly and readnone) can be used in both contexts. Because their interpretation is different in the two contexts and only those two attributes can be used in both contexts, I've opted to keep the Haskell types for parameter and function attributes distinct, but move the two types into separate modules so they can have contructors with the same names.

Documentation

data ParameterAttribute

http://llvm.org/docs/LangRef.html#parameter-attributes

Constructors

ZeroExt

SignExt

InReg

SRet

Alignment Word64

NoAlias

ByVal

NoCapture

Nest

InAlloca

NonNull

Dereferenceable Word64

DereferenceableOrNull Word64

Returned

SwiftSelf

SwiftError

data FunctionAttribute

http://llvm.org/docs/LangRef.html#function-attributes

Constructors

NoReturn

NoUnwind

ReadNone

ReadOnly

NoInline

NoRecurse

AlwaysInline

MinimizeSize

OptimizeForSize

OptimizeNone

StackProtect

StackProtectReq

StackProtectStrong

NoRedZone

NoImplicitFloat

Naked

InlineHint

StackAlignment Word64

ReturnsTwice

UWTable

NonLazyBind

Builtin

NoBuiltin

Cold

JumpTable

NoDuplicate

SanitizeAddress

SanitizeThread

SanitizeMemory

Speculatable

StringAttribute

Fields

stringAttributeKind :: ShortByteString

stringAttributeValue :: ShortByteString

Use "" for no value -- the two are conflated

AllocSize Word32 (Maybe Word32)
AllocSize 0 (Just 0) is invalid

WriteOnly

ArgMemOnly

Convergent

InaccessibleMemOnly

InaccessibleMemOrArgMemOnly

SafeStack

newtype GroupID

http://llvm.org/docs/LangRef.html#attribute-groups

Constructors

GroupID Word