LLVM.AST.Global

Description

Globals - top-level values in Modules - and supporting structures.

Documentation

data Global

http://llvm.org/doxygen/classllvm_1_1GlobalValue.html

Constructors

GlobalVariable
http://llvm.org/docs/LangRef.html#global-variables

Fields

name :: Name

linkage :: Linkage

visibility :: Visibility

dllStorageClass :: Maybe StorageClass

threadLocalMode :: Maybe Model

unnamedAddr :: Maybe UnnamedAddr

isConstant :: Bool

type' :: Type

addrSpace :: AddrSpace

initializer :: Maybe Constant

section :: Maybe ShortByteString

comdat :: Maybe ShortByteString

alignment :: Word32

metadata :: [(ShortByteString, MDRef MDNode)]

GlobalAlias
http://llvm.org/docs/LangRef.html#aliases

Fields

name :: Name

linkage :: Linkage

visibility :: Visibility

dllStorageClass :: Maybe StorageClass

threadLocalMode :: Maybe Model

unnamedAddr :: Maybe UnnamedAddr

type' :: Type

addrSpace :: AddrSpace

aliasee :: Constant

Function
http://llvm.org/docs/LangRef.html#functions

Fields

linkage :: Linkage

visibility :: Visibility

dllStorageClass :: Maybe StorageClass

callingConvention :: CallingConvention

returnAttributes :: [ParameterAttribute]

returnType :: Type

name :: Name

parameters :: ([Parameter], Bool)

snd indicates varargs

functionAttributes :: [Either GroupID FunctionAttribute]

section :: Maybe ShortByteString

comdat :: Maybe ShortByteString

alignment :: Word32

garbageCollectorName :: Maybe ShortByteString

prefix :: Maybe Constant

basicBlocks :: [BasicBlock]

personalityFunction :: Maybe Constant

metadata :: [(ShortByteString, MDRef MDNode)]

data Parameter

Parameters for Functions

Constructors

Parameter Type Name [ParameterAttribute]

data BasicBlock

http://llvm.org/doxygen/classllvm_1_1BasicBlock.html LLVM code in a function is a sequence of BasicBlocks each with a label, some instructions, and a terminator.

Constructors

BasicBlock Name [Named Instruction] (Named Terminator)

data UnnamedAddr

Constructors

LocalAddr

GlobalAddr

globalVariableDefaults :: Global

helper for making GlobalVariables

globalAliasDefaults :: Global

helper for making GlobalAliass

functionDefaults :: Global

helper for making Functions