LLVM.AST.Operand

Description

A type to represent operands to LLVM Instructions

Documentation

data DILocalVariable

https://llvm.org/docs/LangRef.html#dilocalvariable

Constructors

LocalVariable

Fields

name :: ShortByteString

scope :: MDRef DIScope

file :: Maybe (MDRef DIFile)

line :: Word32

type' :: Maybe (MDRef DIType)

flags :: [DIFlag]

arg :: Word16

alignInBits :: Word32

data DIGlobalVariable

https://llvm.org/docs/LangRef.html#diglobalvariable

Constructors

GlobalVariable

Fields

name :: ShortByteString

scope :: Maybe (MDRef DIScope)

file :: Maybe (MDRef DIFile)

line :: Word32

type' :: Maybe (MDRef DIType)

linkageName :: ShortByteString

local :: Bool

definition :: Bool

staticDataMemberDeclaration :: Maybe (MDRef DIDerivedType)

templateParams :: [MDRef DITemplateParameter]

alignInBits :: Word32

data DIVariable

https://llvm.org/doxygen/classllvm_1_1DIVariable.html

Constructors

DIGlobalVariable DIGlobalVariable

DILocalVariable DILocalVariable

data DILexicalBlockBase

https://llvm.org/doxygen/classllvm_1_1DILexicalBlockBase.html

Constructors

DILexicalBlock
https://llvm.org/docs/LangRef.html#dilexicalblock

Fields

scope :: MDRef DILocalScope

file :: Maybe (MDRef DIFile)

line :: Word32

column :: Word16

DILexicalBlockFile
https://llvm.org/docs/LangRef.html#dilexicalblockfile

Fields

scope :: MDRef DILocalScope

file :: Maybe (MDRef DIFile)

discriminator :: Word32

data DITemplateParameter

https://llvm.org/doxygen/classllvm_1_1DITemplateParameter.html

Constructors

DITemplateTypeParameter
https://llvm.org/docs/LangRef.html#ditemplatetypeparameter

Fields

name :: ShortByteString

type' :: Maybe (MDRef DIType)

For DITemplateTypeParameter this field is required, for DITemplateValueParameter it is optional.

DITemplateValueParameter
https://llvm.org/docs/LangRef.html#ditemplatevalueparameter

Fields

name :: ShortByteString

type' :: Maybe (MDRef DIType)

For DITemplateTypeParameter this field is required, for DITemplateValueParameter it is optional.

value :: Maybe Metadata

tag :: TemplateValueParameterTag

data TemplateValueParameterTag

Constructors

TemplateValueParameter

GNUTemplateTemplateParam

GNUTemplateParameterPack

data Encoding

Constructors

AddressEncoding

BooleanEncoding

FloatEncoding

SignedEncoding

SignedCharEncoding

UnsignedEncoding

UnsignedCharEncoding

UTFEncoding

data DICompositeType

https://llvm.org/docs/LangRef.html#dicompositetype

Constructors

DIArrayType

Fields

subscripts :: [DISubrange]

elementTy :: Maybe (MDRef DIType)

sizeInBits :: Word64

alignInBits :: Word32

flags :: [DIFlag]

DIClassType

Fields

scope :: Maybe (MDRef DIScope)

name :: ShortByteString

file :: Maybe (MDRef DIFile)

line :: Word32

flags :: [DIFlag]

derivedFrom :: Maybe (MDRef DIType)

elements :: [MDRef (Either DIDerivedType DISubprogram)]

DIType with tag set to one of Member, Inheritance, Friend or DILocalScope with DISubprogram set to True.

vtableHolder :: Maybe (MDRef DIType)

templateParams :: [DITemplateParameter]

identifier :: ShortByteString

sizeInBits :: Word64

alignInBits :: Word32

DIEnumerationType

Fields

scope :: Maybe (MDRef DIScope)

name :: ShortByteString

file :: Maybe (MDRef DIFile)

line :: Word32

values :: [DIEnumerator]

baseType :: Maybe (MDRef DIType)

identifier :: ShortByteString

sizeInBits :: Word64

alignInBits :: Word32

DIStructureType

Fields

scope :: Maybe (MDRef DIScope)

name :: ShortByteString

file :: Maybe (MDRef DIFile)

line :: Word32

flags :: [DIFlag]

derivedFrom :: Maybe (MDRef DIType)

elements :: [MDRef (Either DIDerivedType DISubprogram)]

DIType with tag set to one of Member, Inheritance, Friend or DILocalScope with DISubprogram set to True.

runtimeLang :: Word16

vtableHolder :: Maybe (MDRef DIType)

identifier :: ShortByteString

sizeInBits :: Word64

alignInBits :: Word32

DIUnionType

Fields

scope :: Maybe (MDRef DIScope)

name :: ShortByteString

file :: Maybe (MDRef DIFile)

line :: Word32

flags :: [DIFlag]

elements :: [MDRef (Either DIDerivedType DISubprogram)]

DIType with tag set to one of Member, Inheritance, Friend or DILocalScope with DISubprogram set to True.

runtimeLang :: Word16

identifier :: ShortByteString

sizeInBits :: Word64

alignInBits :: Word32

data DIDerivedType

https://llvm.org/docs/LangRef.html#diderivedtype

Constructors

DerivedType

Fields

tag :: DerivedTypeTag

name :: ShortByteString

file :: Maybe (MDRef DIFile)

line :: Word32

scope :: Maybe (MDRef DIScope)

baseType :: Maybe (MDRef DIType)

This can be Nothing to represent void *

sizeInBits :: Word64

alignInBits :: Word32

offsetInBits :: Word64

addressSpace :: Maybe Word32

flags :: [DIFlag]

data DerivedTypeTag

Constructors

Typedef

PointerType

PtrToMemberType

ReferenceType

RValueReferenceType

ConstType

VolatileType

RestrictType

AtomicType

Member

Inheritance

Friend

data DISubroutineType

https://llvm.org/docs/LangRef.html#disubroutinetype

Constructors

SubroutineType

Fields

flags :: [DIFlag]

cc :: Word8

typeArray :: [Maybe (MDRef DIType)]

The first element is the return type, the following are the operand types. Nothing corresponds to void.

data DIBasicType

https://llvm.org/docs/LangRef.html#dibasictype

Constructors

BasicType

Fields

name :: ShortByteString

sizeInBits :: Word64

alignInBits :: Word32

encoding :: Maybe Encoding

tag :: BasicTypeTag

flags :: [DIFlag]

data DIType

Constructors

DIBasicType DIBasicType

DICompositeType DICompositeType

DIDerivedType DIDerivedType

DISubroutineType DISubroutineType

data BasicTypeTag

Constructors

BaseType

UnspecifiedType

data Virtuality

Constructors

NoVirtuality

Virtual

PureVirtual

data DISubprogram

https://llvm.org/docs/LangRef.html#disubprogram

Constructors

Subprogram

Fields

scope :: Maybe (MDRef DIScope)

name :: ShortByteString

linkageName :: ShortByteString

file :: Maybe (MDRef DIFile)

line :: Word32

type' :: Maybe (MDRef DISubroutineType)

localToUnit :: Bool

definition :: Bool

scopeLine :: Word32

containingType :: Maybe (MDRef DIType)

virtuality :: Virtuality

virtualityIndex :: Word32

thisAdjustment :: Int32

flags :: [DIFlag]

optimized :: Bool

unit :: Maybe (MDRef DICompileUnit)

templateParams :: [MDRef DITemplateParameter]

declaration :: Maybe (MDRef DISubprogram)

retainedNodes :: [MDRef DILocalVariable]

thrownTypes :: [MDRef DIType]

data DILocalScope

https://llvm.org/doxygen/classllvm_1_1DILocalScope.html

Constructors

DILexicalBlockBase DILexicalBlockBase

DISubprogram DISubprogram

data ChecksumKind

Constructors

MD5

SHA1

data ChecksumInfo

Constructors

ChecksumInfo

Fields

checksumKind :: ChecksumKind

checksumValue :: ShortByteString

data DIFile

https://llvm.org/docs/LangRef.html#difile

Constructors

File

Fields

filename :: ShortByteString

directory :: ShortByteString

checksum :: Maybe ChecksumInfo

data DICompileUnit

https://llvm.org/docs/LangRef.html#dicompileunit

Constructors

CompileUnit

Fields

language :: Word32

file :: MDRef DIFile

producer :: ShortByteString

optimized :: Bool

flags :: ShortByteString

runtimeVersion :: Word32

splitDebugFileName :: ShortByteString

emissionKind :: DebugEmissionKind

enums :: [MDRef DICompositeType]

Only enum types are allowed here

retainedTypes :: [MDRef (Either DIType DISubprogram)]

globals :: [MDRef DIGlobalVariableExpression]

imports :: [MDRef DIImportedEntity]

macros :: [MDRef DIMacroNode]

dWOId :: Word64

splitDebugInlining :: Bool

debugInfoForProfiling :: Bool

nameTableKind :: DebugNameTableKind

debugBaseAddress :: Bool

data DebugNameTableKind

Constructors

NameTableKindDefault

NameTableKindGNU

NameTableKindNone

data DebugEmissionKind

Constructors

NoDebug

FullDebug

LineTablesOnly

data DINamespace

Constructors

Namespace

Fields

name :: ShortByteString

scope :: Maybe (MDRef DIScope)

exportSymbols :: Bool

data DIModule

Constructors

Module

Fields

scope :: Maybe (MDRef DIScope)

name :: ShortByteString

configurationMacros :: ShortByteString

includePath :: ShortByteString

isysRoot :: ShortByteString

data DIScope

https://llvm.org/doxygen/classllvm_1_1DIScope.html

Constructors

DICompileUnit DICompileUnit

DIFile DIFile

DILocalScope DILocalScope

DIModule DIModule

DINamespace DINamespace

DIType DIType

data DICount

Constructors

DICountConstant Int64

DICountVariable (MDRef DIVariable)

data DISubrange

https://llvm.org/docs/LangRef.html#disubrange

Constructors

Subrange

Fields

count :: DICount

lowerBound :: Int64

data DIEnumerator

https://llvm.org/docs/LangRef.html#dienumerator

Constructors

Enumerator

Fields

value :: Int64

isUnsigned :: Bool

name :: ShortByteString

data DIImportedEntity

https://llvm.org/doxygen/classllvm_1_1DIImportedEntity.html

Constructors

ImportedEntity

Fields

tag :: ImportedEntityTag

name :: ShortByteString

scope :: MDRef DIScope

entity :: Maybe (MDRef DINode)

file :: Maybe (MDRef DIFile)

line :: Word32

data ImportedEntityTag

Constructors

ImportedModule

ImportedDeclaration

data DIObjCProperty

https://llvm.org/doxygen/classllvm_1_1DIObjCProperty.html

Constructors

ObjCProperty

Fields

name :: ShortByteString

file :: Maybe (MDRef DIFile)

line :: Word32

getterName :: ShortByteString

setterName :: ShortByteString

attributes :: Word32

type' :: Maybe (MDRef DIType)

data DINode

https://llvm.org/doxygen/classllvm_1_1DINode.html

Constructors

DIEnumerator DIEnumerator

DIImportedEntity DIImportedEntity

DIObjCProperty DIObjCProperty

DIScope DIScope

DISubrange DISubrange

DITemplateParameter DITemplateParameter

DIVariable DIVariable

data DIMacroNode

https://llvm.org/doxygen/classllvm_1_1DIMacroNode.html

Constructors

DIMacro
https://llvm.org/docs/LangRef.html#dimacro

Fields

info :: DIMacroInfo

line :: Word32

name :: ShortByteString

value :: ShortByteString

DIMacroFile
https://llvm.org/docs/LangRef.html#dimacrofile

Fields

line :: Word32

file :: MDRef DIFile

elements :: [MDRef DIMacroNode]

data DIMacroInfo

Constructors

Define

Undef

data DIFlag

Constructors

Accessibility DIAccessibility

FwdDecl

AppleBlock

BlockByrefStruct

VirtualFlag

Artificial

Explicit

Prototyped

ObjcClassComplete

ObjectPointer

Vector

StaticMember

LValueReference

RValueReference

InheritanceFlag DIInheritance

IntroducedVirtual

BitField

NoReturn

MainSubprogram

data DIInheritance

Inheritance flag

Constructors

SingleInheritance

MultipleInheritance

VirtualInheritance

data DIAccessibility

Accessiblity flag

Constructors

Private

Protected

Public

data DIGlobalVariableExpression

A pair of a DIVariable and a MDNode.

This is used in the cuGlobals fields of DIScope.

Constructors

GlobalVariableExpression

Fields

var :: MDRef DIGlobalVariable

expr :: MDRef DIExpression

data DIExpression

https://llvm.org/docs/LangRef.html#diexpression

Constructors

Expression

Fields

operands :: [DWOp]

data DILocation

Constructors

Location

Fields

line :: Word32

column :: Word16

scope :: MDRef DILocalScope

data MDNode

http://llvm.org/docs/LangRef.html#metadata

Constructors

MDTuple [Maybe Metadata]
Nothing represents null

DIExpression DIExpression

DIGlobalVariableExpression DIGlobalVariableExpression

DILocation DILocation

DIMacroNode DIMacroNode

DINode DINode

data DWOp

https://llvm.org/docs/LangRef.html#diexpression

Constructors

DwOpFragment DWOpFragment
Must appear at the end

DW_OP_StackValue
Must be the last one or followed by a DW_OP_LLVM_Fragment

DW_OP_Swap

DW_OP_ConstU Word64

DW_OP_Lit0

DW_OP_PlusUConst Word64

DW_OP_Plus

DW_OP_Minus

DW_OP_Mul

DW_OP_Div

DW_OP_Mod

DW_OP_Not

DW_OP_Or

DW_OP_Xor

DW_OP_And

DW_OP_Shr

DW_OP_Shra

DW_OP_Shl

DW_OP_Dup

DW_OP_Deref

DW_OP_XDeref

data DWOpFragment

Constructors

DW_OP_LLVM_Fragment

Fields

offset :: Word64

size :: Word64

data MDRef a

MDRef can either represent a reference to some piece of metadata or the metadata itself.

This is mainly useful for encoding cyclic metadata. Note that LLVM represents inline and non-inline nodes identically, so roundtripping the Haskell AST does not preserve whether a node was inline or not.

Constructors

MDRef MetadataNodeID

MDInline a

newtype MetadataNodeID

A MetadataNodeID is a number for identifying a metadata node. Note this is different from "named metadata", which are represented with NamedMetadataDefinition.

Constructors

MetadataNodeID Word

data Metadata

http://llvm.org/docs/LangRef.html#metadata

Constructors

MDString ShortByteString
http://llvm.org/docs/doxygen/html/classllvm_1_1MDNode.html

MDNode (MDRef MDNode)
http://llvm.org/docs/doxygen/html/classllvm_1_1MDNode.html

MDValue Operand
http://llvm.org/docs/doxygen/html/classllvm_1_1ValueAsMetadata.html

type CallableOperand = Either InlineAssembly Operand

The Call instruction is special: the callee can be inline assembly

data Operand

An Operand is roughly that which is an argument to an Instruction

Constructors

LocalReference Type Name
%foo

ConstantOperand Constant
Constants include GlobalReference, for @foo

MetadataOperand Metadata