LLVM.IRBuilder.Instruction

Documentation

fadd :: MonadIRBuilder m => Operand -> Operand -> m Operand

See reference.

fmul :: MonadIRBuilder m => Operand -> Operand -> m Operand

See reference.

fsub :: MonadIRBuilder m => Operand -> Operand -> m Operand

See reference.

fdiv :: MonadIRBuilder m => Operand -> Operand -> m Operand

See reference.

frem :: MonadIRBuilder m => Operand -> Operand -> m Operand

See reference.

add :: MonadIRBuilder m => Operand -> Operand -> m Operand

See reference.

mul :: MonadIRBuilder m => Operand -> Operand -> m Operand

See reference.

sub :: MonadIRBuilder m => Operand -> Operand -> m Operand

See reference.

udiv :: MonadIRBuilder m => Operand -> Operand -> m Operand

See reference.

sdiv :: MonadIRBuilder m => Operand -> Operand -> m Operand

See reference.

urem :: MonadIRBuilder m => Operand -> Operand -> m Operand

See reference.

srem :: MonadIRBuilder m => Operand -> Operand -> m Operand

See reference.

shl :: MonadIRBuilder m => Operand -> Operand -> m Operand

See reference.

lshr :: MonadIRBuilder m => Operand -> Operand -> m Operand

See reference.

ashr :: MonadIRBuilder m => Operand -> Operand -> m Operand

See reference.

and :: MonadIRBuilder m => Operand -> Operand -> m Operand

See reference.

or :: MonadIRBuilder m => Operand -> Operand -> m Operand

See reference.

xor :: MonadIRBuilder m => Operand -> Operand -> m Operand

See reference.

alloca :: MonadIRBuilder m => Type -> Maybe Operand -> Word32 -> m Operand

See reference.

load :: MonadIRBuilder m => Operand -> Word32 -> m Operand

See reference.

store :: MonadIRBuilder m => Operand -> Word32 -> Operand -> m ()

See reference.

gep :: (MonadIRBuilder m, MonadModuleBuilder m) => Operand -> [Operand] -> m Operand

Emit the getelementptr instruction. See reference.

trunc :: MonadIRBuilder m => Operand -> Type -> m Operand

Emit the trunc ... to instruction. See reference.

fptrunc :: MonadIRBuilder m => Operand -> Type -> m Operand

Emit the fptrunc ... to instruction. See reference.

zext :: MonadIRBuilder m => Operand -> Type -> m Operand

Emit the zext ... to instruction. See reference.

sext :: MonadIRBuilder m => Operand -> Type -> m Operand

Emit the sext ... to instruction. See reference.

fptoui :: MonadIRBuilder m => Operand -> Type -> m Operand

Emit the fptoui ... to instruction. See reference.

fptosi :: MonadIRBuilder m => Operand -> Type -> m Operand

Emit the fptosi ... to instruction. See reference.

fpext :: MonadIRBuilder m => Operand -> Type -> m Operand

Emit the fpext ... to instruction. See reference.

uitofp :: MonadIRBuilder m => Operand -> Type -> m Operand

Emit the uitofp ... to instruction. See reference.

sitofp :: MonadIRBuilder m => Operand -> Type -> m Operand

Emit the sitofp ... to instruction. See reference.

ptrtoint :: MonadIRBuilder m => Operand -> Type -> m Operand

Emit the ptrtoint ... to instruction. See reference.

inttoptr :: MonadIRBuilder m => Operand -> Type -> m Operand

Emit the inttoptr ... to instruction. See reference.

bitcast :: MonadIRBuilder m => Operand -> Type -> m Operand

Emit the bitcast ... to instruction. See reference.

extractElement :: MonadIRBuilder m => Operand -> Operand -> m Operand

See reference.

insertElement :: MonadIRBuilder m => Operand -> Operand -> Operand -> m Operand

See reference.

shuffleVector :: MonadIRBuilder m => Operand -> Operand -> Constant -> m Operand

See reference.

extractValue :: MonadIRBuilder m => Operand -> [Word32] -> m Operand

See reference.

insertValue :: MonadIRBuilder m => Operand -> Operand -> [Word32] -> m Operand

See reference.

icmp :: MonadIRBuilder m => IntegerPredicate -> Operand -> Operand -> m Operand

See reference.

fcmp :: MonadIRBuilder m => FloatingPointPredicate -> Operand -> Operand -> m Operand

See reference.

br :: MonadIRBuilder m => Name -> m ()

Unconditional branch. Emit a br label dest instruction See reference.

phi :: MonadIRBuilder m => [(Operand, Name)] -> m Operand

See reference.

retVoid :: MonadIRBuilder m => m ()

Emit a ret void instruction. See reference.

call :: MonadIRBuilder m => Operand -> [(Operand, [ParameterAttribute])] -> m Operand

See reference.

ret :: MonadIRBuilder m => Operand -> m ()

See reference.

switch :: MonadIRBuilder m => Operand -> Name -> [(Constant, Name)] -> m ()

See reference.

select :: MonadIRBuilder m => Operand -> Operand -> Operand -> m Operand

See reference.

condBr :: MonadIRBuilder m => Operand -> Name -> Name -> m ()

Conditional branch (see br for unconditional instructions). See reference.

unreachable :: MonadIRBuilder m => m ()

See reference.

globalStringPtr

Arguments

:: (MonadModuleBuilder m, MonadIRBuilder m)

=> String
The string to generate

-> Name
Variable name of the pointer

-> m Operand

Creates a series of instructions to generate a pointer to a string constant. Useful for making format strings to pass to printf, for example