LLVM.IRBuilder.Monad

Documentation

newtype IRBuilderT m a

This provides a uniform API for creating instructions and inserting them into a basic block: either at the end of a BasicBlock, or at a specific location in a block.

Constructors

IRBuilderT

Fields

unIRBuilderT :: StateT IRBuilderState m a

type IRBuilder = IRBuilderT Identity

class Monad m => MonadIRBuilder m where

Minimal complete definition

Nothing

Methods

liftIRState :: State IRBuilderState a -> m a

liftIRState :: (MonadTrans t, MonadIRBuilder m1, m ~ t m1) => State IRBuilderState a -> m a

data PartialBlock

A partially constructed block as a sequence of instructions

Constructors

PartialBlock

Fields

partialBlockName :: !Name

partialBlockInstrs :: SnocList (Named Instruction)

partialBlockTerm :: Maybe (Named Terminator)

emptyPartialBlock :: Name -> PartialBlock

data IRBuilderState

Builder monad state

Constructors

IRBuilderState

Fields

builderSupply :: !Word

builderUsedNames :: !(HashSet ShortByteString)

builderNameSuggestion :: !(Maybe ShortByteString)

builderBlocks :: SnocList BasicBlock

builderBlock :: !(Maybe PartialBlock)

emptyIRBuilder :: IRBuilderState

runIRBuilder :: IRBuilderState -> IRBuilder a -> (a, [BasicBlock])

Evaluate IRBuilder to a result and a list of basic blocks

runIRBuilderT :: Monad m => IRBuilderState -> IRBuilderT m a -> m (a, [BasicBlock])

Evaluate IRBuilderT to a result and a list of basic blocks

execIRBuilder :: IRBuilderState -> IRBuilder a -> [BasicBlock]

Evaluate IRBuilder to a list of basic blocks

execIRBuilderT :: Monad m => IRBuilderState -> IRBuilderT m a -> m [BasicBlock]

Evaluate IRBuilderT to a list of basic blocks

Low-level functionality

ensureBlock :: MonadIRBuilder m => m ()

If no partial block exists, create a new block with a fresh label.

This is useful if you want to ensure that the label for the block is assigned before another label which is not possible with modifyBlock.

modifyBlock :: MonadIRBuilder m => (PartialBlock -> PartialBlock) -> m ()

fresh :: MonadIRBuilder m => m Name

Generate a fresh name. The resulting name is numbered or based on the name suggested with named if that's used.

freshName :: MonadIRBuilder m => ShortByteString -> m Name

Generate a fresh name from a name suggestion

freshUnName :: MonadIRBuilder m => m Name

Generate a fresh numbered name

emitInstr

Arguments

:: MonadIRBuilder m

=> Type
Return type

-> Instruction

-> m Operand

Emit instruction

emitInstrVoid :: MonadIRBuilder m => Instruction -> m ()

Emit instruction that returns void

emitTerm :: MonadIRBuilder m => Terminator -> m ()

Emit terminator

emitBlockStart :: MonadIRBuilder m => Name -> m ()

Starts a new block labelled using the given name and ends the previous one. The name is assumed to be fresh.

High-level functionality

block :: MonadIRBuilder m => m Name

Starts a new block and ends the previous one

named :: MonadIRBuilder m => m r -> ShortByteString -> m r

ir named name executes the IRBuilder ir using name as the base name whenever a fresh local name is generated. Collisions are avoided by appending numbers (first "name", then "name1", "name2", and so on).

currentBlock :: MonadIRBuilder m => m Name

Get the name of the currently active block.

This function will throw an error if there is no active block. The only situation in which this can occur is if it is called before any call to block and before emitting any instructions.

hasTerminator :: MonadIRBuilder m => m Bool

Find out if the currently active block has a terminator.

This function will fail under the same condition as currentBlock