LLVM.TH

Description

This module is presents template haskell mostly like the template-haskell >= 2.10 / ghc >= 7.10, even on earlier versions. It's intended as an internal library for llvm-hs-pure and llvm-hs; it's exposed only to be shared between the two.

Documentation

module Language.Haskell.TH

conT :: Typish qt => Name -> qt

appT :: Typish qt => qt -> Q Type -> qt