LLVM.AST

Description

This module and descendants define AST data types to represent LLVM code. Note that these types are designed for fidelity rather than convenience - if the truth of what LLVM supports is less than pretty, so be it.

Documentation

data Module

http://llvm.org/docs/LangRef.html#module-structure

Constructors

Module

Fields

moduleName :: ShortByteString

moduleSourceFileName :: ShortByteString

moduleDataLayout :: Maybe DataLayout

a DataLayout, if specified, must match that of the eventual code generator

moduleTargetTriple :: Maybe ShortByteString

moduleDefinitions :: [Definition]

defaultModule :: Module

helper for making Modules

data Definition

Any thing which can be at the top level of a Module

Constructors

GlobalDefinition Global

TypeDefinition Name (Maybe Type)

MetadataNodeDefinition MetadataNodeID [Maybe Metadata]

NamedMetadataDefinition ShortByteString [MetadataNodeID]

ModuleInlineAssembly ByteString

FunctionAttributes GroupID [FunctionAttribute]

COMDAT ShortByteString SelectionKind

data Global

http://llvm.org/doxygen/classllvm_1_1GlobalValue.html

Constructors

GlobalVariable Name Linkage Visibility (Maybe StorageClass) (Maybe Model) (Maybe UnnamedAddr) Bool Type AddrSpace (Maybe Constant) (Maybe ShortByteString) (Maybe ShortByteString) Word32
http://llvm.org/docs/LangRef.html#global-variables

GlobalAlias Name Linkage Visibility (Maybe StorageClass) (Maybe Model) (Maybe UnnamedAddr) Type AddrSpace Constant
http://llvm.org/docs/LangRef.html#aliases

Function Linkage Visibility (Maybe StorageClass) CallingConvention [ParameterAttribute] Type Name ([Parameter], Bool) [Either GroupID FunctionAttribute] (Maybe ShortByteString) (Maybe ShortByteString) Word32 (Maybe ShortByteString) (Maybe Constant) [BasicBlock] (Maybe Constant)
http://llvm.org/docs/LangRef.html#functions

globalVariableDefaults :: Global

helper for making GlobalVariables

globalAliasDefaults :: Global

helper for making GlobalAliass

functionDefaults :: Global

helper for making Functions

data UnnamedAddr

Constructors

LocalAddr

GlobalAddr

data Parameter

Parameters for Functions

Constructors

Parameter Type Name [ParameterAttribute]

data BasicBlock

http://llvm.org/doxygen/classllvm_1_1BasicBlock.html LLVM code in a function is a sequence of BasicBlocks each with a label, some instructions, and a terminator.

Constructors

BasicBlock Name [Named Instruction] (Named Terminator)

module LLVM.AST.Instruction

module LLVM.AST.Name

module LLVM.AST.Operand

module LLVM.AST.Type