LLVM.AST.Instruction

Description

LLVM instructions http://llvm.org/docs/LangRef.html#instruction-reference

Documentation

type InstructionMetadata = [(ShortByteString, MetadataNode)]

http://llvm.org/docs/LangRef.html#metadata-nodes-and-metadata-strings Metadata can be attached to an instruction

data Terminator

http://llvm.org/docs/LangRef.html#terminators

Constructors

Ret

Fields

returnOperand :: Maybe Operand

metadata' :: InstructionMetadata

CondBr

Fields

condition :: Operand

trueDest :: Name

falseDest :: Name

metadata' :: InstructionMetadata

Br

Fields

dest :: Name

metadata' :: InstructionMetadata

Switch

Fields

operand0' :: Operand

defaultDest :: Name

dests :: [(Constant, Name)]

metadata' :: InstructionMetadata

IndirectBr

Fields

operand0' :: Operand

possibleDests :: [Name]

metadata' :: InstructionMetadata

Invoke

Fields

callingConvention' :: CallingConvention

returnAttributes' :: [ParameterAttribute]

function' :: CallableOperand

arguments' :: [(Operand, [ParameterAttribute])]

functionAttributes' :: [Either GroupID FunctionAttribute]

returnDest :: Name

exceptionDest :: Name

metadata' :: InstructionMetadata

Resume

Fields

operand0' :: Operand

metadata' :: InstructionMetadata

Unreachable

Fields

metadata' :: InstructionMetadata

CleanupRet

Fields

cleanupPad :: Operand

unwindDest :: Maybe Name

metadata' :: InstructionMetadata

CatchRet

Fields

catchPad :: Operand

successor :: Name

metadata' :: InstructionMetadata

CatchSwitch

Fields

parentPad' :: Operand

catchHandlers :: NonEmpty Name

defaultUnwindDest :: Maybe Name

metadata' :: InstructionMetadata

data FastMathFlags

http://llvm.org/docs/LangRef.html#fast-math-flags

Constructors

NoFastMathFlags

UnsafeAlgebra

FastMathFlags

Fields

noNaNs :: Bool

noInfs :: Bool

noSignedZeros :: Bool

allowReciprocal :: Bool

data MemoryOrdering

http://llvm.org/docs/LangRef.html#atomic-memory-ordering-constraints http://llvm.org/docs/Atomics.html

Constructors

Unordered

Monotonic

Acquire

Release

AcquireRelease

SequentiallyConsistent

data SynchronizationScope

http://llvm.org/docs/LangRef.html#singlethread

Constructors

SingleThread

System

type Atomicity = (SynchronizationScope, MemoryOrdering)

An Atomicity describes constraints on the visibility of effects of an atomic instruction

data LandingPadClause

For the redoubtably complex LandingPad instruction

Constructors

Catch Constant

Filter Constant

data TailCallKind

For the call instruction http://llvm.org/docs/LangRef.html#call-instruction

Constructors

Tail

MustTail

data Instruction

non-terminator instructions: http://llvm.org/docs/LangRef.html#binaryops http://llvm.org/docs/LangRef.html#bitwiseops http://llvm.org/docs/LangRef.html#memoryops http://llvm.org/docs/LangRef.html#otherops

Constructors

Add

Fields

nsw :: Bool

nuw :: Bool

operand0 :: Operand

operand1 :: Operand

metadata :: InstructionMetadata

FAdd

Fields

fastMathFlags :: FastMathFlags

operand0 :: Operand

operand1 :: Operand

metadata :: InstructionMetadata

Sub

Fields

nsw :: Bool

nuw :: Bool

operand0 :: Operand

operand1 :: Operand

metadata :: InstructionMetadata

FSub

Fields

fastMathFlags :: FastMathFlags

operand0 :: Operand

operand1 :: Operand

metadata :: InstructionMetadata

Mul

Fields

nsw :: Bool

nuw :: Bool

operand0 :: Operand

operand1 :: Operand

metadata :: InstructionMetadata

FMul

Fields

fastMathFlags :: FastMathFlags

operand0 :: Operand

operand1 :: Operand

metadata :: InstructionMetadata

UDiv

Fields

exact :: Bool

operand0 :: Operand

operand1 :: Operand

metadata :: InstructionMetadata

SDiv

Fields

exact :: Bool

operand0 :: Operand

operand1 :: Operand

metadata :: InstructionMetadata

FDiv

Fields

fastMathFlags :: FastMathFlags

operand0 :: Operand

operand1 :: Operand

metadata :: InstructionMetadata

URem

Fields

operand0 :: Operand

operand1 :: Operand

metadata :: InstructionMetadata

SRem

Fields

operand0 :: Operand

operand1 :: Operand

metadata :: InstructionMetadata

FRem

Fields

fastMathFlags :: FastMathFlags

operand0 :: Operand

operand1 :: Operand

metadata :: InstructionMetadata

Shl

Fields

nsw :: Bool

nuw :: Bool

operand0 :: Operand

operand1 :: Operand

metadata :: InstructionMetadata

LShr

Fields

exact :: Bool

operand0 :: Operand

operand1 :: Operand

metadata :: InstructionMetadata

AShr

Fields

exact :: Bool

operand0 :: Operand

operand1 :: Operand

metadata :: InstructionMetadata

And

Fields

operand0 :: Operand

operand1 :: Operand

metadata :: InstructionMetadata

Or

Fields

operand0 :: Operand

operand1 :: Operand

metadata :: InstructionMetadata

Xor

Fields

operand0 :: Operand

operand1 :: Operand

metadata :: InstructionMetadata

Alloca

Fields

allocatedType :: Type

numElements :: Maybe Operand

alignment :: Word32

metadata :: InstructionMetadata

Load

Fields

volatile :: Bool

address :: Operand

maybeAtomicity :: Maybe Atomicity

alignment :: Word32

metadata :: InstructionMetadata

Store

Fields

volatile :: Bool

address :: Operand

value :: Operand

maybeAtomicity :: Maybe Atomicity

alignment :: Word32

metadata :: InstructionMetadata

GetElementPtr

Fields

inBounds :: Bool

address :: Operand

indices :: [Operand]

metadata :: InstructionMetadata

Fence

Fields

atomicity :: Atomicity

metadata :: InstructionMetadata

CmpXchg

Fields

volatile :: Bool

address :: Operand

expected :: Operand

replacement :: Operand

atomicity :: Atomicity

failureMemoryOrdering :: MemoryOrdering

metadata :: InstructionMetadata

AtomicRMW

Fields

volatile :: Bool

rmwOperation :: RMWOperation

address :: Operand

value :: Operand

atomicity :: Atomicity

metadata :: InstructionMetadata

Trunc

Fields

operand0 :: Operand

type' :: Type

metadata :: InstructionMetadata

ZExt

Fields

operand0 :: Operand

type' :: Type

metadata :: InstructionMetadata

SExt

Fields

operand0 :: Operand

type' :: Type

metadata :: InstructionMetadata

FPToUI

Fields

operand0 :: Operand

type' :: Type

metadata :: InstructionMetadata

FPToSI

Fields

operand0 :: Operand

type' :: Type

metadata :: InstructionMetadata

UIToFP

Fields

operand0 :: Operand

type' :: Type

metadata :: InstructionMetadata

SIToFP

Fields

operand0 :: Operand

type' :: Type

metadata :: InstructionMetadata

FPTrunc

Fields

operand0 :: Operand

type' :: Type

metadata :: InstructionMetadata

FPExt

Fields

operand0 :: Operand

type' :: Type

metadata :: InstructionMetadata

PtrToInt

Fields

operand0 :: Operand

type' :: Type

metadata :: InstructionMetadata

IntToPtr

Fields

operand0 :: Operand

type' :: Type

metadata :: InstructionMetadata

BitCast

Fields

operand0 :: Operand

type' :: Type

metadata :: InstructionMetadata

AddrSpaceCast

Fields

operand0 :: Operand

type' :: Type

metadata :: InstructionMetadata

ICmp

Fields

iPredicate :: IntegerPredicate

operand0 :: Operand

operand1 :: Operand

metadata :: InstructionMetadata

FCmp

Fields

fpPredicate :: FloatingPointPredicate

operand0 :: Operand

operand1 :: Operand

metadata :: InstructionMetadata

Phi

Fields

type' :: Type

incomingValues :: [(Operand, Name)]

metadata :: InstructionMetadata

Call

Fields

tailCallKind :: Maybe TailCallKind

callingConvention :: CallingConvention

returnAttributes :: [ParameterAttribute]

function :: CallableOperand

arguments :: [(Operand, [ParameterAttribute])]

functionAttributes :: [Either GroupID FunctionAttribute]

metadata :: InstructionMetadata

Select

Fields

condition' :: Operand

trueValue :: Operand

falseValue :: Operand

metadata :: InstructionMetadata

VAArg

Fields

argList :: Operand

type' :: Type

metadata :: InstructionMetadata

ExtractElement

Fields

vector :: Operand

index :: Operand

metadata :: InstructionMetadata

InsertElement

Fields

vector :: Operand

element :: Operand

index :: Operand

metadata :: InstructionMetadata

ShuffleVector

Fields

operand0 :: Operand

operand1 :: Operand

mask :: Constant

metadata :: InstructionMetadata

ExtractValue

Fields

aggregate :: Operand

indices' :: [Word32]

metadata :: InstructionMetadata

InsertValue

Fields

aggregate :: Operand

element :: Operand

indices' :: [Word32]

metadata :: InstructionMetadata

LandingPad

Fields

type' :: Type

cleanup :: Bool

clauses :: [LandingPadClause]

metadata :: InstructionMetadata

CatchPad

Fields

catchSwitch :: Operand

args :: [Operand]

metadata :: InstructionMetadata

CleanupPad

Fields

parentPad :: Operand

args :: [Operand]

metadata :: InstructionMetadata

data Named a

Instances of instructions may be given a name, allowing their results to be referenced as Operands. Sometimes instructions - e.g. a call to a function returning void - don't need names.

Constructors

Name := a

Do a