LLVM.IRBuilder.Module

Documentation

newtype ModuleBuilderT m a

Constructors

ModuleBuilderT

Fields

unModuleBuilderT :: StateT ModuleBuilderState m a

newtype ModuleBuilderState

Constructors

ModuleBuilderState

Fields

builderDefs :: SnocList Definition

emptyModuleBuilder :: ModuleBuilderState

type ModuleBuilder = ModuleBuilderT Identity

class Monad m => MonadModuleBuilder m where

Methods

liftModuleState :: State ModuleBuilderState a -> m a

liftModuleState :: (MonadTrans t, MonadModuleBuilder m1, m ~ t m1) => State ModuleBuilderState a -> m a

runModuleBuilder :: ModuleBuilderState -> ModuleBuilder a -> (a, [Definition])

Evaluate ModuleBuilder to a result and a list of definitions

runModuleBuilderT :: Monad m => ModuleBuilderState -> ModuleBuilderT m a -> m (a, [Definition])

Evaluate ModuleBuilderT to a result and a list of definitions

execModuleBuilder :: ModuleBuilderState -> ModuleBuilder a -> [Definition]

Evaluate ModuleBuilder to a list of definitions

execModuleBuilderT :: Monad m => ModuleBuilderState -> ModuleBuilderT m a -> m [Definition]

Evaluate ModuleBuilderT to a list of definitions

emitDefn :: MonadModuleBuilder m => Definition -> m ()

data ParameterName

A parameter name suggestion

Constructors

NoParameterName

ParameterName ShortByteString

function

Arguments

:: MonadModuleBuilder m

=> Name
Function name

-> [(Type, ParameterName)]
Parameter types and name suggestions

-> Type
Return type

-> ([Operand] -> IRBuilderT m ())
Function body builder

-> m Operand

Define and emit a (non-variadic) function definition

extern

Arguments

:: MonadModuleBuilder m

=> Name
Definition name

-> [Type]
Parametere types

-> Type
Type

-> m Operand

An external function definition

typedef :: MonadModuleBuilder m => Name -> Maybe Type -> m ()

A named type definition

buildModule :: ShortByteString -> ModuleBuilder a -> Module

Convenience function for module construction

buildModuleT :: Monad m => ShortByteString -> ModuleBuilderT m a -> m Module

Convenience function for module construction (transformer version)